# congratulator-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "congratulator",
    "contact": {
      "name": "ASTiNEZh",
      "url": "https://t.me/ASTiNEZh",
      "email": "astinezh@gmail.com"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "локальный сервер для разработки"
    }
  ],
  "paths": {
    "/birthdays": {
      "get": {
        "tags": [
          "BirthdaysController"
        ],
        "summary": "Отобразить все имеющиеся ДР",
        "operationId": "ReturnAllBirthdays",
        "responses": {
          "200": {
            "description": "Передача данных клиенту",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllBirthdays"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ResponseMessage"
          }
        }
      }
    },
    "/birthdays/upcoming": {
      "get": {
        "tags": [
          "BirthdaysController"
        ],
        "summary": "Отобразить все наступающие ДР",
        "operationId": "ReturnActuallBirthdays",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "задать интересующий период отчётности",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 366,
              "default": 31
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/schemas/ActualBirthdays"
          },
          "default": {
            "$ref": "#/components/responses/ResponseMessage"
          }
        }
      }
    },
    "/birthdays/new": {
      "post": {
        "tags": [
          "BirthdaysController"
        ],
        "summary": "Создать новый ДР",
        "operationId": "CreateNewBirthday",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Birthday"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseMessage"
          },
          "default": {
            "$ref": "#/components/responses/ResponseMessage"
          }
        }
      }
    },
    "/birthdays/сorrecting/{birthdayId}": {
      "patch": {
        "tags": [
          "BirthdaysController"
        ],
        "summary": "Редактировать данные о ДР",
        "operationId": "CorrectNewBirthday",
        "parameters": [
          {
            "name": "birthdayId",
            "in": "path",
            "description": "задать интересующий период отчётности",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Birthday"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseMessage"
          },
          "default": {
            "$ref": "#/components/responses/ResponseMessage"
          }
        }
      }
    },
    "/birthdays/delete/{birthdayId}": {
      "delete": {
        "tags": [
          "BirthdaysController"
        ],
        "summary": "Отредактировать данные о ДР",
        "operationId": "DeleteNewBirthday",
        "parameters": [
          {
            "name": "birthdayId",
            "in": "path",
            "description": "задать интересующий период отчётности",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseMessage"
          },
          "default": {
            "$ref": "#/components/responses/ResponseMessage"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "ResponseMessage": {
        "description": "Результат выполнения запроса",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Похоже на сервере произошёл сбой. Ваша проблема уже решается.",
                  "description": "Описание результата выполнения запроса / причина ошибки"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "Birthday": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid"
          },
          "lastName": {
            "type": "string",
            "description": "Фамилия"
          },
          "firstName": {
            "type": "string",
            "description": "Имя"
          },
          "middleName": {
            "type": "string",
            "description": "Отчество"
          },
          "birthdayDate": {
            "type": "string",
            "format": "date-time",
            "description": "Дата дня рождения"
          }
        }
      },
      "AllBirthdays": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Birthday"
        }
      },
      "ActualBirthdays": {
        "type": "object",
        "properties": {
          "todayBirthdays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Birthday"
            },
            "description": "Список сегоднешних именинников"
          },
          "upcomingBirthdays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Birthday"
            },
            "description": "Список наступающих именин"
          }
        }
      }
    }
  }
}